Implementing clean architecture
