+++ src/WebAPI/.gitignore
@@ -15,5 +15,4 @@
 # along with sep3cs. If not, see <http://www.gnu.org/licenses/>.
 #
 
-*
-!.gitignore
+package-lock.json
