+++ src/WebAPI/ClientApp/aspnetcore-https.js
