+++ src/WebAPI/ClientApp/aspnetcore-react.js
