+++ src/WebAPI/ClientApp/public/favicon.ico
