+++ src/WebAPI/ClientApp/public/manifest.json
