+++ src/WebAPI/ClientApp/src/components/ErrorDialog.js
