+++ src/WebAPI/ClientApp/src/service-worker.js
