+++ src/WebAPI/ClientApp/src/serviceWorkerRegistration.js
