+++ src/WebAPI/Program.cs
@@ -15,9 +15,28 @@
  * along with sep3cs. If not, see <http://www.gnu.org/licenses/>.
  */
 
-namespace DataClash.Data
+namespace DataClash
 {
-  public class Player : User
+  class Program
     {
+      public static int Main (string[] args)
+        {
+          var builder = WebApplication.CreateBuilder (args);
+
+          builder.Services.AddControllersWithViews ();
+
+          var app = builder.Build ();
+
+          if (app.Environment.IsDevelopment () == false)
+            app.UseHsts ();
+
+          app.UseHttpsRedirection ();
+          app.UseStaticFiles ();
+          app.UseRouting ();
+          app.MapControllerRoute (name : "default", pattern : "{controller}/{action=Index}/{id?}");
+          app.MapFallbackToFile ("index.html");
+          app.Run ();
+        return 0;
+        }
     }
 }
